A tool to mock out REST responses for given data
We want to be able to mock out responses for data, so that
we can test many cases (nested data structures, errors, etc).

This uses wiremock to mock out the REST server, and tools for
us to queue up responses to be served.  That way we can test
how we handle such responses.

This is sort of awkwardly placed; the dirtree should be improved.