Extract execution connection functions to ClientConnection
PrestoClient was serving two roles -- one as the main entrypoint to
submit statements for execution, and the other as the object which
actively managed the connection to the server.  This was a bit
confusing, because some of its methods should only have been invoked
in the first role, and some in the second. Separating them out
makes it a bit cleaner, which each struct having a clearer role.